fix(platform): typo
